Update patch set 5
Patch Set 5:

`Components defined at the root level of your project, in the components folder, are global and can be used anywhere in your application. But if you decide to define a new component inside another component (nesting), this new component can only be used its direct parent.`

The text above says that the components nested can only be used by his parent component in our case is the Card component. In the scenes, the Card component just use the CardHeader and CardBody.
My point in putting CardHeader and CardBody inside the folder Card was to avoid that we create a CardHeader/CardBody without a Card.

If we put the way you suggested:
class Card extends component {

render() {
  const  {title, marginLeft, children, editable} = this.prop
  return () {
    <CardHeader>
      <CardTitle editable title={title} marginLeft={marginLeft} />
    </CardHeader>
    <CardBody>
       {children}
    </CardBody>
  }
}
} export default Card

We cannot anymore "extend" the CardHeader or have a CardTitle that is not editable.

Patch-set: 5
Reviewer: Ramon Henrique <1013096@d5d70762-12d0-45a1-890d-524b12d3f735>